app/result-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.4.2
prometheus-fastapi-instrumentator==6.1.0
pytest==7.4.3
